+++ btx/processing/tasks/calculate_p_values.py
@@ -4,7 +4,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 import warnings
 
-from btx.processing.types import CalculatePValuesInput, CalculatePValuesOutput
+from btx.processing.btx_types import CalculatePValuesInput, CalculatePValuesOutput
 
 class CalculatePValues:
     """Calculate p-values from EMD values and null distribution."""
